Added Video & User back-end

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 // database.find({형식}, 콜백함수 )
 // Video.find({}, (error, videos) => {}); 
@@ -14,7 +15,7 @@ export const home = async(req, res) => {
 // render(name of view, variable)
 export const watch = async (req, res) => {
     const id = req.params.id;    //const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.render("404", {pageTitle:"Video not found"});        
     } 
@@ -22,20 +23,29 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);        
     if(video === null){
         return res.status(404).render("404", {pageTitle:"Video not found."})        
     } 
+
+    //owner만 수정가능
+    if(String(video.owner) !== String(_id))
+        return res.status(403).redirect("/");
     return res.render("edit", {pageTitle:`Edit: ${video.title}`, video});
     
 } ;
+
 export const postEdit = async (req, res) =>{
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id:id}); // true or false   
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
+    if(String(video.owner) !== String(_id))
+        return res.status(403).redirect("/");
     await Video.findByIdAndUpdate(id, {
         title, description, hashtags:Video.formatHashtags(hashtags)
     }); 
@@ -45,6 +55,7 @@ export const getUpload = (req, res) => {
   return res.render("upload", {pageTitle : "Upload Video"});  
 };
 export const postUpload = async (req, res) => {   
+    const {user: {_id}} = req.session;
     const file = req.file; 
     const {title, hashtags, description} = req.body; 
         
@@ -63,12 +74,16 @@ export const postUpload = async (req, res) => {
     // const dbVideo = await video.save();
     
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title, // title: title
             description,            
             fileUrl: file.path,
+            owner:_id,
             hashtags: Video.formatHashtags(hashtags)            
         }); 
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);        
+        user.save();
         return res.redirect("/");
     } catch(error){
         return res.status(400).render("upload", {pageTitle : "Upload Video", errorMessage: error._message}); 
@@ -77,6 +92,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+
+    //validation
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id))
+        return res.status(403).redirect("/");
+
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
